Refactor Http/2 Implementation
Complete refactor of the http/2 implementation as the previous version did
not work for requests bodies and did not properly encapsulate the multipart
request object. This refactor provides a functional http/2 service that
should for for any verb type. One call out is that is makes the assumption
that no requests will have `Content-Length: 0`. This may or may not be a bad
idea.