Add Http/2 Server (#21)
* Add Http/2 Server

This is a basic implementation of Http/2 which will auto-negotiate and
should return the same value for Http/1.1 and Http/2 requests. This commit
does not handle anything in the request body as of yet but is a good
starting point.

* Refactor Http/2 Implementation

Complete refactor of the http/2 implementation as the previous version did
not work for requests bodies and did not properly encapsulate the multipart
request object. This refactor provides a functional http/2 service that
should for for any verb type. One call out is that is makes the assumption
that no requests will have `Content-Length: 0`. This may or may not be a bad
idea.

* Refactor Http/2 Implementation

Complete refactor of the http/2 implementation as the previous version did
not work for requests bodies and did not properly encapsulate the multipart
request object. This refactor provides a functional http/2 service that
should for for any verb type. One call out is that is makes the assumption
that no requests will have `Content-Length: 0`. This may or may not be a bad
idea.

* Modifications to reflect code review updates

In this commit a few things have been changed or updated. One of the main
additions was an ordered AtomicReference for the routes list to handle concurrent
access and updates to the routes. The IOException was also added back for
the Handler class.