Initial Http/1.1 Client Implementation
This is the initial Implementation with the base client. It includes retries
as well as an async call stack. Currently, the test requires the Example
class to be running, but that should be stubbed out to make the unit test
self contained. Also, the ALPN code is still included but commented out
until I can get the Http/2 impl completed. Please disregard that section of
code for now.