Ulr and pipeline fixes (#35)
* UrlRouter will match on path only instead of full URL

Also, added tests for the XUrl helper class. Additionally, added @Getter for
both the server channel (which allows access to the pipeline) as well as for
the worker eventLoopGroup which will allow for tasks to be executed in the
worker thread pool.

* Added XHttpMethod to allow binding multiple handlers to the same path based
on Http Method type.

* Added XHttpMethod to allow binding multiple handlers to the same path based
on Http Method type.