Return 429 on too many requests and fix npe (#51)
* Return 429 on too many requests and fix npe

This commit addresses issues 45 and 50.

For Issue 45:
2 new limits were added to the config, a "soft" limit and a "hard" limit.
Once the soft limit is reached, the server will return 429s directly. If the
hard limit is reached, the server will block on the worker thread to protect
the back end from being overwhelmed and thus leading to resource exhaustion.

For Issue 50:
The healthCheckRegistry was be scheduled to run prior to the worker group
being properly bootstrapped. This commit changes the order in which those
processes are run. This was added to this commit for convenience.

* address PR comments and replace infer with spotbugs

* remove erroneous comments

* addresses open PR comments