Enhancement/license management (#100)
* Add Apache-2.0 template, add license plugin

* Minor config changes in license plugin. Update license headers

* Sets the Java mapping to SLASHSTAR_STYLE to be consistent with what we have
  already.
* Adds missing license headers, updates license headers
* Misc. googleJavaFormat updates

* Removing gradle-license-plugin

There are at least a few bugs in this plugin that make me want to remove it
until they can be addressed in a PR or by the maintainer.

There is at least one issue open in the project related to this:

* https://github.com/hierynomus/license-gradle-plugin/issues/14
  * Changing the substitution variables doesn't update as expected
  (i.e. changing the year)

Running `./gradle licenseFormat` appears to "fix" a number of files with missing
headers according to it's output, but there are no changes to files listed. Also
running `./gradlew licenseMain` or `./gradlew check` fails to find the license headers.